backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
boto3 = "^1.34.92"
django-storages = {extras = ["s3"], version = "^1.14.2"}
pypdf2 = "^3.0.1"
tiktoken = "^0.6.0"
faiss-cpu = "1.7.4"
jinja2 = "3.1.2"
altair = "4.2.2"
pinecone-client = "^3.2.2"
pytesseract = "^0.3.10"
tesseract = "^0.1.3"
langchain = "^0.1.16"
langchain-pinecone = "^0.1.0"
openai = "^1.23.6"
langchain-openai = "^0.1.4"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
